requirements.txt
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0
django-pdfkit
django-widget-tweaks==1.4.12
django-wkhtmltopdf==3.4.0
wkhtmltopdf-pack-ng
gunicorn
html5lib==1.1
psycopg2==2.9.3
python-bidi==0.4.2
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
dj_database_url